+++ src/lib/ProjectM36/StaticOptimizer.hs
@@ -17,6 +17,7 @@ import ProjectM36.NormalizeExpr
 import qualified ProjectM36.Attribute as A
 import qualified ProjectM36.AttributeNames as AS
 import ProjectM36.Streaming.RelationalExpression
+import Control.Monad (void)
 #if MIN_VERSION_base(4,18,0)
 import Control.Monad (foldM)
 #endif