+++ test/TransactionGraph/Automerge.hs
@@ -79,7 +79,7 @@ testAutomergeFailure = TestCase $ do
 testAutomergeReconnect :: Test
 testAutomergeReconnect = TestCase $ withSystemTempDirectory "m36testdb" $ \tempdir -> do
   let repro = do
-          conn <- unsafeLeftCrash =<< connectProjectM36 (InProcessConnectionInfo (CrashSafePersistence (tempdir </> "test.db")) emptyNotificationCallback [] basicDatabaseContext)
+          conn <- unsafeLeftCrash =<< connectProjectM36 (InProcessConnectionInfo (CrashSafePersistence (tempdir </> "test.db")) emptyNotificationCallback [] basicDatabaseContext [superAdminRole])
           sess <- unsafeLeftCrash =<< createSessionAtHead conn "master"
           autoMergeToHead sess conn UnionMergeStrategy "master"
         -- commit sess conn